Fix #136726: Cycles: Area light visible when size is 0
Area light with a size of zero should not contribute to the scene, so
set the light as disabled.

This does not only fix the reported bug where such light is visible to
the camera, but also a regression in 4.2 where the light contributes to
the scene when light tree is off.

Pull Request: https://projects.blender.org/blender/blender/pulls/136763