Modeling: "Snap to Cursor" now supports "Rotation"
Adds an operator setting to allow matching the rotation of the snapped
object to the rotation of the 3d cursor.

When the operators "Rotation" option is enabled, it will rotate selected
objects to match the current rotation of the 3d cursor.
Rotation modes and axis locking are respected.

If multiple objects are selected, they will all be set to the rotation
of the 3d cursor. There's an opportunity for future work on this for
the case where "offset" is toggled on as well, so that the location and
rotation of offset objects would be determined by rotating around the
pivot point.

Addresses #134863.

Ref: !134963